FrSky Sport: refactor packing code into common library
This should allow it to be reused for UAVOTaranis and also
perhaps the FrSky Sensor Hub code (with a different packing
constant).
