+++ flight/Modules/UAVOFrSKYSPortBridge/UAVOFrSKYSPortBridge.c
@@ -33,77 +33,19 @@
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-#include "pios.h"
-#include "openpilot.h"
-#include "modulesettings.h"
-#include "misc_math.h"
-#include "physical_constants.h"
-#include "attitudeactual.h"
+#include "frsky_packing.h"
+#include "pios_thread.h"
+
 #include "baroaltitude.h"
-#include "positionactual.h"
-#include "velocityactual.h"
 #include "flightbatterysettings.h"
 #include "flightbatterystate.h"
 #include "gpsposition.h"
-#include "gpstime.h"
-#include "homelocation.h"
-#include "accels.h"
-#include "flightstatus.h"
-#include "airspeedactual.h"
-#include "nedaccel.h"
-#include "velocityactual.h"
-#include "attitudeactual.h"
-#include "pios_thread.h"
+#include "modulesettings.h"
 
 #if defined(PIOS_INCLUDE_FRSKY_SPORT_TELEMETRY)
 
-static void uavoFrSKYSPortBridgeTask(void *parameters);
-static bool frsky_encode_gps_course(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_altitude(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_vario(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_current(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_cells(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_t1(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_t2(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_fuel(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_acc(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_gps_coord(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_gps_alt(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_gps_speed(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_gps_time(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_rpm(uint32_t *value, bool test_presence_only, uint32_t arg);
-static bool frsky_encode_airspeed(uint32_t *value, bool test_presence_only, uint32_t arg);
-
 #define FRSKY_POLL_REQUEST                 0x7e
 #define FRSKY_MINIMUM_POLL_INTERVAL        10000
-
-enum frsky_value_id {
-	FRSKY_ALT_ID = 0x0100,
-	FRSKY_VARIO_ID = 0x110,
-	FRSKY_CURR_ID = 0x0200,
-	FRSKY_VFAS_ID = 0x0210,
-	FRSKY_CELLS_ID = 0x0300,
-	FRSKY_T1_ID = 0x0400,
-	FRSKY_T2_ID = 0x0410,
-	FRSKY_RPM_ID = 0x0500,
-	FRSKY_FUEL_ID = 0x0600,
-	FRSKY_ACCX_ID = 0x0700,
-	FRSKY_ACCY_ID = 0x0710,
-	FRSKY_ACCZ_ID = 0x0720,
-	FRSKY_GPS_LON_LAT_ID = 0x0800,
-	FRSKY_GPS_ALT_ID = 0x0820,
-	FRSKY_GPS_SPEED_ID = 0x0830,
-	FRSKY_GPS_COURSE_ID = 0x0840,
-	FRSKY_GPS_TIME_ID = 0x0850,
-	FRSKY_ADC3_ID = 0x0900,
-	FRSKY_ADC4_ID = 0x0910,
-	FRSKY_AIR_SPEED_ID = 0x0a00,
-	FRSKY_RSSI_ID = 0xf101,
-	FRSKY_ADC1_ID = 0xf102,
-	FRSKY_ADC2_ID = 0xf103,
-	FRSKY_BATT_ID = 0xf104,
-	FRSKY_SWR_ID = 0xf105,
-};
 
 enum frsky_state {
 	FRSKY_STATE_WAIT_POLL_REQUEST,
@@ -111,13 +53,7 @@ enum frsky_state {
 	FRSKY_STATE_WAIT_TX_DONE,
 };
 
-struct frsky_value_item {
-	enum frsky_value_id id;
-	uint16_t period_ms;
-	bool (*encode_value)(uint32_t *value, bool test_presence_only, uint32_t arg);
-	uint32_t fn_arg;
-};
-
+// Set of objects sent by this module
 static const struct frsky_value_item frsky_value_items[] = {
 	{FRSKY_GPS_COURSE_ID,  100,   frsky_encode_gps_course, 0}, // attitude yaw estimate
 	{FRSKY_ALT_ID,         100,   frsky_encode_altitude,   0}, // altitude estimate
@@ -145,21 +81,17 @@ static const struct frsky_value_item frsky_value_items[] = {
 	{FRSKY_AIR_SPEED_ID,   100,   frsky_encode_airspeed,   0}, // airspeed
 };
 
-static const uint8_t frsky_sensor_ids[] = {0x1b};
 struct frsky_sport_telemetry {
-	struct pios_thread *task;
-	uintptr_t com;
-	uint32_t item_last_triggered[NELEMENTS(frsky_value_items)];
-	uint32_t last_poll_time;
-	uint8_t ignore_rx_chars;
 	enum frsky_state state;
 	int32_t scheduled_item;
-	bool use_current_sensor;
-	uint8_t batt_cell_count;
-	bool use_baro_sensor;
-	FlightBatterySettingsData battery_settings;
-	GPSPositionData gps_position;
+	uint32_t last_poll_time;
+	uint8_t ignore_rx_chars;
+	uintptr_t com;
+	struct frsky_settings frsky_settings;
+	uint32_t item_last_triggered[NELEMENTS(frsky_value_items)];
 };
+
+static const uint8_t frsky_sensor_ids[] = {0x1b};
 
 #define FRSKY_SPORT_BAUDRATE                    57600
 
@@ -172,569 +104,8 @@ struct frsky_sport_telemetry {
 
 static bool module_enabled;
 static struct frsky_sport_telemetry *frsky;
-
-/**
- * Encode baro altitude value
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_altitude(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (!frsky->use_baro_sensor || (PositionActualHandle() == NULL))
-		return false;
-
-	if (test_presence_only)
-		return true;
-	// instead of encoding baro altitude directly, we will use
-	// more accurate estimation in PositionActual UAVO
-	float down = 0;
-	PositionActualDownGet(&down);
-	int32_t alt = (int32_t)(-down * 100.0f);
-	*value = (uint32_t) alt;
-	return true;
-}
-
-/**
- * Encode heading value
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_gps_course(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (AttitudeActualHandle() == NULL)
-		return false;
-
-	if (test_presence_only)
-		return true;
-
-	AttitudeActualData attitude;
-	AttitudeActualGet(&attitude);
-	float hdg = (attitude.Yaw >= 0) ? attitude.Yaw : (attitude.Yaw + 360.0f);
-	*value = (uint32_t)(hdg * 100.0f);
-
-	return true;
-}
-
-/**
- * Encode vertical speed value
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_vario(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (!frsky->use_baro_sensor || VelocityActualHandle() == NULL)
-		return false;
-
-	if (test_presence_only)
-		return true;
-
-	float down = 0;
-	VelocityActualDownGet(&down);
-	int32_t vspeed = (int32_t)(-down * 100.0f);
-	*value = (uint32_t) vspeed;
-
-	return true;
-}
-
-/**
- * Encode battery current value
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_current(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (!frsky->use_current_sensor)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	float current = 0;
-	FlightBatteryStateCurrentGet(&current);
-	int32_t current_frsky = (int32_t)(current * 10.0f);
-	*value = (uint32_t) current_frsky;
-
-	return true;
-}
-
-/**
- * Encode battery cells voltage
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[], index of battery cell pair
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_cells(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if ((frsky->batt_cell_count == 0) || (frsky->batt_cell_count - 1) < (arg * 2))
-		return false;
-	if (test_presence_only)
-		return true;
-
-	float voltage = 0;
-	FlightBatteryStateVoltageGet(&voltage);
-
-	uint32_t cell_voltage = (uint32_t)((voltage * 500.0f) / frsky->batt_cell_count);
-	*value = ((cell_voltage & 0xfff) << 8) | ((arg * 2) & 0x0f) | ((frsky->batt_cell_count << 4) & 0xf0);
-	if (((int16_t)frsky->batt_cell_count - 1) >= (arg * 2 + 1))
-		*value |= ((cell_voltage & 0xfff) << 20);
-
-	return true;
-}
-
-/**
- * Encode GPS status as T1 value
- * Right-most two digits means visible satellite count, left-most digit has following meaning:
- * 1 - no GPS connected
- * 2 - no fix
- * 3 - 2D fix
- * 4 - 3D fix
- * 5 - 3D fix and HomeLocation is SET - should be safe for navigation
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value successfully encoded or presence test passed
- */
-static bool frsky_encode_t1(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL)
-		return false;
-	
-	if (test_presence_only)
-		return true;
-	
-	uint8_t hl_set = HOMELOCATION_SET_FALSE;
-	
-	if (HomeLocationHandle())
-		HomeLocationSetGet(&hl_set);
-
-	int32_t t1 = 0;
-	switch (frsky->gps_position.Status) {
-	case GPSPOSITION_STATUS_NOGPS:
-		t1 = 100;
-		break;
-	case GPSPOSITION_STATUS_NOFIX:
-		t1 = 200;
-		break;
-	case GPSPOSITION_STATUS_FIX2D:
-		t1 = 300;
-		break;
-	case GPSPOSITION_STATUS_FIX3D:
-	case GPSPOSITION_STATUS_DIFF3D:
-		if (hl_set == HOMELOCATION_SET_TRUE)
-			t1 = 500;
-		else
-			t1 = 400;
-		break;
-	}
-	if (frsky->gps_position.Satellites > 0)
-		t1 += frsky->gps_position.Satellites;
-
-	*value = (uint32_t)t1;
-
-	return true;
-}
-
-/**
- * Encode GPS hDop and vDop as T2
- * Bits 0-7  = hDop * 100, max 255 (hDop = 2.55)
- * Bits 8-15 = vDop * 100, max 255 (vDop = 2.55)
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value successfully encoded or presence test passed
- */
-static bool frsky_encode_t2(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL)
-		return false;
-
-	if (test_presence_only)
-		return true;
-
-	uint32_t hdop = (uint32_t)(frsky->gps_position.HDOP * 100.0f);
-
-	if (hdop > 255)
-		hdop = 255;
-			
-	uint32_t vdop = (uint32_t)(frsky->gps_position.VDOP * 100.0f);
-			
-	if (vdop > 255)
-		vdop = 255;
-	
-	*value = 256 * vdop + hdop;
-
-	return true;
-}
-
-/**
- * Encode consumed battery energy as fuel
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_fuel(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (!frsky->use_current_sensor)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	uint32_t capacity = frsky->battery_settings.Capacity;
-	float consumed_mahs = 0;
-	FlightBatteryStateConsumedEnergyGet(&consumed_mahs);
-
-	float fuel =  (uint32_t)(100.0f * (1.0f - consumed_mahs / capacity));
-	fuel = bound_min_max(fuel, 0.0f, 100.0f);
-	*value = (uint32_t) fuel;
-
-	return true;
-}
-
-/**
- * Encode configured values
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]; 0=X, 1=Y, 2=Z
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_acc(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	uint8_t accelDataSettings;
-	ModuleSettingsFrskyAccelDataGet(&accelDataSettings);
-
-	float acc = 0;
-	switch(accelDataSettings) {
-	case MODULESETTINGS_FRSKYACCELDATA_ACCELS: {
-		if (AccelsHandle() == NULL)
-			return false;
-		else if (test_presence_only)
-			return true;
-
-		switch (arg) {
-		case 0:
-			AccelsxGet(&acc);
-			break;
-		case 1:
-			AccelsyGet(&acc);
-			break;
-		case 2:
-			AccelszGet(&acc);
-			break;
-		}
-
-		acc /= GRAVITY;
-		acc *= 100.0f;
-		break;
-	}
-
-	case MODULESETTINGS_FRSKYACCELDATA_NEDACCELS: {
-		if (NedAccelHandle() == NULL)
-			return false;
-		else if (test_presence_only)
-			return true;
-
-		switch (arg) {
-		case 0:
-			NedAccelNorthGet(&acc);
-			break;
-		case 1:
-			NedAccelEastGet(&acc);
-			break;
-		case 2:
-			NedAccelDownGet(&acc);
-			break;
-		}
-
-		acc /= GRAVITY;
-		acc *= 100.0f;
-		break;
-	}
-
-	case MODULESETTINGS_FRSKYACCELDATA_NEDVELOCITY: {
-		if (VelocityActualHandle() == NULL)
-			return false;
-		else if (test_presence_only)
-			return true;
-
-		switch (arg) {
-		case 0:
-			VelocityActualNorthGet(&acc);
-			break;
-		case 1:
-			VelocityActualEastGet(&acc);
-			break;
-		case 2:
-			VelocityActualDownGet(&acc);
-			break;
-		}
-
-		acc *= 100.0f;
-		break;
-	}
-
-	case MODULESETTINGS_FRSKYACCELDATA_ATTITUDEANGLES: {
-		if (AttitudeActualHandle() == NULL)
-			return false;
-		else if (test_presence_only)
-			return true;
-
-		switch (arg) {
-		case 0:
-			AttitudeActualRollGet(&acc);
-			break;
-		case 1:
-			AttitudeActualPitchGet(&acc);
-			break;
-		case 2:
-			AttitudeActualYawGet(&acc);
-			break;
-		}
-
-		acc *= 100.0f;
-		break;
-	}
-	}
-
-	int32_t frsky_acc = (int32_t) acc;
-	*value = (uint32_t) frsky_acc;
-
-	return true;
-}
-
-/**
- * Encode gps coordinates
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]; 0=lattitude, 1=longitude
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_gps_coord(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL)
-		return false;
-	if (frsky->gps_position.Status == GPSPOSITION_STATUS_NOFIX
-			|| frsky->gps_position.Status == GPSPOSITION_STATUS_NOGPS)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	uint32_t frsky_coord = 0;
-	int32_t coord = 0;
-	if (arg == 0) {
-		// lattitude
-		coord = frsky->gps_position.Latitude;
-		if (coord >= 0)
-			frsky_coord = 0;
-		else
-			frsky_coord = 1 << 30;
-	} else {
-		// longitude
-		coord = frsky->gps_position.Longitude;
-		if (coord >= 0)
-			frsky_coord = 2 << 30;
-		else
-			frsky_coord = 3 << 30;
-	}
-	coord = abs(coord);
-	frsky_coord |= (((uint64_t)coord * 6ull) / 100);
-
-	*value = frsky_coord;
-	return true;
-}
-
-/**
- * Encode gps altitude
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_gps_alt(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL)
-		return false;
-	if (frsky->gps_position.Status != GPSPOSITION_STATUS_FIX3D
-			&& frsky->gps_position.Status != GPSPOSITION_STATUS_DIFF3D)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	int32_t frsky_gps_alt = (int32_t)(frsky->gps_position.Altitude * 100.0f);
-	*value = (uint32_t)frsky_gps_alt;
-
-	return true;
-}
-
-/**
- * Encode gps speed
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_gps_speed(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL)
-		return false;
-	if (frsky->gps_position.Status != GPSPOSITION_STATUS_FIX3D
-			&& frsky->gps_position.Status != GPSPOSITION_STATUS_DIFF3D)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	int32_t frsky_speed = (int32_t)((frsky->gps_position.Groundspeed / KNOTS2M_PER_SECOND) * 1000);
-	*value = frsky_speed;
-	return true;
-}
-
-/**
- * Encode GPS UTC time
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]; 0=date, 1=time
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_gps_time(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (GPSPositionHandle() == NULL || GPSTimeHandle() == NULL)
-		return false;
-	if (frsky->gps_position.Status != GPSPOSITION_STATUS_FIX3D
-			&& frsky->gps_position.Status != GPSPOSITION_STATUS_DIFF3D)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	GPSTimeData gps_time;
-	GPSTimeGet(&gps_time);
-	uint32_t frsky_time = 0;
-
-	if (arg == 0) {
-		// encode date
-		frsky_time = 0x000000ff;
-		frsky_time |= gps_time.Day << 8;
-		frsky_time |= gps_time.Month << 16;
-		frsky_time |= (gps_time.Year % 100) << 24;
-	} else {
-		frsky_time = 0;
-		frsky_time |= gps_time.Second << 8;
-		frsky_time |= gps_time.Minute << 16;
-		frsky_time |= gps_time.Hour << 24;
-	}
-	*value = frsky_time;
-	return true;
-}
-
-/**
- * Encodes ARM status and flight mode number as RPM value
- * Since there is no RPM information in any UAVO available,
- * we will intentionaly misuse this item to encode another useful information.
- * It will encode flight status as three-digit number as follow:
- * most left digit encodes arm status (1=armed, 0=disarmed)
- * two most right digits encodes flight mode number (see FlightStatus UAVO FlightMode enum)
- * To work this propperly on Taranis, you have to set Blades to "1" in telemetry setting
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_rpm(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (FlightStatusHandle() == NULL)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	FlightStatusData flight_status;
-	FlightStatusGet(&flight_status);
-
-	*value = (flight_status.Armed == FLIGHTSTATUS_ARMED_ARMED) ? 200 : 100;
-	*value += flight_status.FlightMode;
-
-	return true;
-}
-
-/**
- * Encode true airspeed(TAS)
- * @param[out] value encoded value
- * @param[in] test_presence_only true when function should only test for availability of this value
- * @param[in] arg argument specified in frsky_value_items[]
- * @returns true when value succesfully encoded or presence test passed
- */
-static bool frsky_encode_airspeed(uint32_t *value, bool test_presence_only, uint32_t arg)
-{
-	if (AirspeedActualHandle() == NULL)
-		return false;
-	if (test_presence_only)
-		return true;
-
-	AirspeedActualData airspeed;
-	AirspeedActualGet(&airspeed);
-	int32_t frsky_speed = (int32_t)((airspeed.TrueAirspeed / KNOTS2M_PER_SECOND) * 10);
-	*value = (uint32_t)frsky_speed;
-
-	return true;
-}
-
-/**
- * Performs byte stuffing and checksum calculation
- * @param[out] obuff buffer where byte stuffed data will came in
- * @param[in,out] chk checksum byte to update
- * @param[in] byte
- * @returns count of bytes inserted to obuff (1 or 2)
- */
-static uint8_t frsky_insert_byte(uint8_t *obuff, uint16_t *chk, uint8_t byte)
-{
-	/* checksum calculation is based on data before byte-stuffing */
-	*chk += byte;
-	*chk += (*chk) >> 8;
-	*chk &= 0x00ff;
-	*chk += (*chk) >> 8;
-	*chk &= 0x00ff;
-
-	if (byte == 0x7e || byte == 0x7d) {
-		obuff[0] = 0x7d;
-		obuff[1] = byte &= ~0x20;
-		return 2;
-	}
-
-	obuff[0] = byte;
-	return 1;
-}
-/**
- * Send u32 value dataframe to FrSky SmartPort bus
- * @param[in] id FrSky value ID
- * @param[in] value value
- */
-static void frsky_send_frame(enum frsky_value_id id, uint32_t value)
-{
-	/* each call of frsky_insert_byte can add 2 bytes to the buffer at maximum
-	 * and therefore the worst-case is 15 bytes total (the first byte 0x10 won't be
-	 * escaped) */
-	uint8_t tx_data[15];
-	uint16_t chk = 0;
-	uint8_t cnt = 0;
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, 0x10);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, (uint16_t)id & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, ((uint16_t)id >> 8) & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, value & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, (value >> 8) & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, (value >> 16) & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, (value >> 24) & 0xff);
-	cnt += frsky_insert_byte(&tx_data[cnt], &chk, 0xff - chk);
-
-	PIOS_COM_SendBuffer(frsky->com, tx_data, cnt);
-
-	frsky->ignore_rx_chars = cnt;
-}
+static int32_t uavoFrSKYSPortBridgeInitialize(void);
+static void uavoFrSKYSPortBridgeTask(void *parameters);
 
 /**
  * Scan for value item with the longest expired time and schedule it to send in next poll turn
@@ -746,7 +117,7 @@ static void frsky_schedule_next_item(void)
 	int32_t max_exp_time = INT32_MIN;
 	int32_t most_exp_item = -1;
 	for (i = 0; i < NELEMENTS(frsky_value_items); i++) {
-		if (frsky_value_items[i].encode_value(0, true, frsky_value_items[i].fn_arg)) {
+		if (frsky_value_items[i].encode_value(&frsky->frsky_settings, 0, true, frsky_value_items[i].fn_arg)) {
 			int32_t exp_time = PIOS_DELAY_GetuSSince(frsky->item_last_triggered[i]) -
 					(frsky_value_items[i].period_ms * 1000);
 			if (exp_time > max_exp_time) {
@@ -767,9 +138,9 @@ static bool frsky_send_scheduled_item(void)
 	if ((item >= 0) && (item < NELEMENTS(frsky_value_items))) {
 		frsky->item_last_triggered[item] = PIOS_DELAY_GetuS();
 		uint32_t value = 0;
-		if (frsky_value_items[item].encode_value(&value, false,
+		if (frsky_value_items[item].encode_value(&frsky->frsky_settings, &value, false,
 				frsky_value_items[item].fn_arg)) {
-			frsky_send_frame((uint16_t)(frsky_value_items[item].id), value);
+			frsky->ignore_rx_chars += frsky_send_frame(frsky->com, (uint16_t)(frsky_value_items[item].id), value);
 			return true;
 		}
 	}
@@ -809,7 +180,7 @@ static void frsky_receive_byte(uint8_t b)
 				// get GPSPositionData once here to be more efficient, since
 				// GPS position data are very often used by encode() handlers
 				if (GPSPositionHandle() != NULL)
-					GPSPositionGet(&frsky->gps_position);
+					GPSPositionGet(&frsky->frsky_settings.gps_position);
 				// send item previously scheduled
 				if (frsky_send_scheduled_item() && frsky->ignore_rx_chars)
 					frsky->state = FRSKY_STATE_WAIT_TX_DONE;
@@ -836,19 +207,20 @@ static int32_t uavoFrSKYSPortBridgeStart(void)
 		uint8_t currentPin;
 		FlightBatterySettingsCurrentPinGet(&currentPin);
 		if (currentPin != FLIGHTBATTERYSETTINGS_CURRENTPIN_NONE)
-			frsky->use_current_sensor = true;
-		FlightBatterySettingsGet(&frsky->battery_settings);
-		frsky->batt_cell_count = frsky->battery_settings.NbCells;
+			frsky->frsky_settings.use_current_sensor = true;
+		FlightBatterySettingsGet(&frsky->frsky_settings.battery_settings);
+		frsky->frsky_settings.batt_cell_count = frsky->frsky_settings.battery_settings.NbCells;
 	}
 	if (BaroAltitudeHandle() != NULL
 			&& PIOS_SENSORS_GetQueue(PIOS_SENSOR_BARO) != NULL)
-		frsky->use_baro_sensor = true;
+		frsky->frsky_settings.use_baro_sensor = true;
 
-	frsky->task = PIOS_Thread_Create(
+	struct pios_thread *task;
+	task = PIOS_Thread_Create(
 			uavoFrSKYSPortBridgeTask, "uavoFrSKYSPortBridge",
 			STACK_SIZE_BYTES, NULL, TASK_PRIORITY);
 	TaskMonitorAdd(TASKINFO_RUNNING_UAVOFRSKYSPORTBRIDGE,
-			frsky->task);
+			task);
 
 	return 0;
 }
@@ -870,14 +242,14 @@ static int32_t uavoFrSKYSPortBridgeInitialize(void)
 		if (frsky != NULL) {
 			memset(frsky, 0x00, sizeof(struct frsky_sport_telemetry));
 
-			frsky->com = sport_com;
-			frsky->ignore_rx_chars = 0;
-			frsky->last_poll_time = PIOS_DELAY_GetuS();
+			frsky->frsky_settings.use_current_sensor = false;
+			frsky->frsky_settings.batt_cell_count = 0;
+			frsky->frsky_settings.use_baro_sensor = false;
 			frsky->state = FRSKY_STATE_WAIT_POLL_REQUEST;
+			frsky->last_poll_time = PIOS_DELAY_GetuS();
+			frsky->ignore_rx_chars = 0;
 			frsky->scheduled_item = -1;
-			frsky->use_current_sensor = false;
-			frsky->batt_cell_count = 0;
-			frsky->use_baro_sensor = false;
+			frsky->com = sport_com;
 
 			uint8_t i;
 			for (i = 0; i < NELEMENTS(frsky_value_items); i++)
